src/projectService/requirements.txt
grpc==1.0.0
grpcio==1.51.1
grpclib==0.4.3
protobuf==4.21.12
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.3.20
